refactor: enums package의 파일들을 domain으로 이동
