+++ src/main/kotlin/blackjack/domain/Action.kt
@@ -1,4 +1,4 @@
-package blackjack.enums
+package blackjack.domain
 
 enum class Action {
     HIT,
