+++ src/main/kotlin/blackjack/domain/Card.kt
@@ -1,8 +1,5 @@
 package blackjack.domain
 
-import blackjack.enums.Rank
-import blackjack.enums.Suit
-
 class Card(
     val rank: Rank,
     val suit: Suit,
