+++ src/main/kotlin/blackjack/domain/Player.kt
@@ -1,7 +1,5 @@
 package blackjack.domain
 
-import blackjack.enums.Result
-
 class Player(
     val name: String,
 ) : Participant() {
