+++ src/main/kotlin/blackjack/domain/Rank.kt
@@ -1,4 +1,4 @@
-package blackjack.enums
+package blackjack.domain
 
 enum class Rank(
     val number: Int,
