+++ src/main/kotlin/blackjack/domain/Result.kt
@@ -1,4 +1,4 @@
-package blackjack.enums
+package blackjack.domain
 
 enum class Result(
     val message: String,
